Match run-command and description in docs
Some requested PR changes caused the example commands and their
descriptions to get out of sync (long vs. short options).  Long options
are easier for beginners.

Setting the workdir to `/data` on the command line is redundant; the
option is dropped from the examples to reduce the amount of information
necessary to use the image.